pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rf_to_dlc"
version = "1.0.0"
description = "Convert Roboflow annotated datasets to DeepLabCut format"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your@email.com"}
]
keywords = ["roboflow", "deeplabcut", "pose-estimation", "keypoints", "coco"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "opencv-python>=4.5.0",
    "matplotlib>=3.4.0",
    "requests>=2.25.0",
    "Pillow>=8.0.0",
    "tqdm>=4.60.0",
    "tables>=3.6.0",
]

[project.optional-dependencies]
gcs = ["google-cloud-storage>=1.35.0"]
dlc = ["deeplabcut>=2.3.0"]
dev = ["pytest", "black", "ruff"]

[project.scripts]
rf-to-dlc = "rf_to_dlc.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/rf_to_dlc"

[tool.setuptools.packages.find]
where = ["."]
include = ["rf_to_dlc*"]
